tasks module is done

+++ src/common/sw_pow.cc
@@ -1,5 +1,5 @@
-int sw_pow(int value, int power) noexcept {
-	int result = value;
+double sw_pow(double value, int power) noexcept {
+	double result = value;
 
 	if (power == 0) return 1;
 	for (int i = 1; i < power; i++) result *= value;
